Add first user data access tests
