+++ lib/galaxy/webapps/galaxy/controllers/page.py
@@ -13,9 +13,9 @@ from galaxy.managers.pages import (
     PageManager,
 )
 from galaxy.managers.sharable import SlugBuilder
-from galaxy.managers.users import get_user_by_username
 from galaxy.managers.workflows import WorkflowsManager
 from galaxy.model.base import transaction
+from galaxy.model.db.user import get_user_by_username
 from galaxy.model.item_attrs import UsesItemRatings
 from galaxy.schema.schema import CreatePagePayload
 from galaxy.structured_app import StructuredApp
