+++ lib/tool_shed/webapp/controllers/repository.py
@@ -24,8 +24,8 @@ from galaxy import (
     util,
     web,
 )
-from galaxy.managers.users import get_user_by_username
 from galaxy.model.base import transaction
+from galaxy.model.db.user import get_user_by_username
 from galaxy.tool_shed.util import dependency_display
 from galaxy.tools.repositories import ValidationContext
 from galaxy.util.tool_shed import encoding_util
